Added previews for every part of App. Separated screens to different files
