+++ app/src/main/java/ru/voodster/composeweather/compose/WeatherApp.kt
@@ -1,16 +1,26 @@
-package ru.voodster.composeweather
+package ru.voodster.composeweather.compose
 
 import androidx.annotation.StringRes
-import androidx.compose.animation.animateColorAsState
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
 import androidx.compose.material.*
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.DateRange
 import androidx.compose.material.icons.filled.Home
 import androidx.compose.material.icons.filled.List
-import androidx.compose.runtime.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.SideEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import androidx.navigation.NavDestination
 import androidx.navigation.NavGraph
 import androidx.navigation.NavHostController
@@ -18,11 +28,12 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.google.accompanist.insets.ProvideWindowInsets
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
-import ru.voodster.composeweather.compose.MainDestinations
-
-import ru.voodster.composeweather.compose.WeatherNavGraph
-import ru.voodster.composeweather.ui.theme.*
-
+import ru.voodster.composeweather.R
+import ru.voodster.composeweather.WeatherViewModel
+import ru.voodster.composeweather.ui.theme.ComposeWeatherTheme
+import ru.voodster.composeweather.ui.theme.primaryDarkColor
+import ru.voodster.composeweather.ui.theme.primaryLightColor
+import ru.voodster.composeweather.ui.theme.primaryTextColor
 
 
 private val NavGraph.startDestination: NavDestination?
@@ -38,10 +49,10 @@ private tailrec fun findStartDestination(graph: NavDestination): NavDestination 
 fun WeatherApp(
     appContainer: WeatherViewModel
 ) {
-    ComposeWeatherTheme() { //
+    ComposeWeatherTheme { //
         ProvideWindowInsets(windowInsetsAnimationsEnabled = true) { // обьявляем обработку вставок типа клавиатуры или navBar
             val systemUiController = rememberSystemUiController() // Контроллер системного UI -
-                                                        // теже клавиатура navBar statusBar
+            // теже клавиатура navBar statusBar
             SideEffect {
                 systemUiController.isStatusBarVisible = false
                 //systemUiController.setSystemBarsColor(Color.Transparent, darkIcons = false)
@@ -56,13 +67,14 @@ fun WeatherApp(
             val bottomNavigationItems = listOf(
                 BottomNavigationScreens.CURRENT,
                 BottomNavigationScreens.TABLE,
-                BottomNavigationScreens.CHART)
+                BottomNavigationScreens.CHART
+            )
 
             Scaffold(
                 scaffoldState = scaffoldState,
-                bottomBar = {BottomNavigationBar(navController, bottomNavigationItems)}
+                bottomBar = { BottomNavigationBar(navController, bottomNavigationItems) }
             )
-            {innerPadding->
+            { innerPadding ->
                 WeatherNavGraph(
                     viewModel = appContainer,
                     navController = navController,
@@ -75,7 +87,7 @@ fun WeatherApp(
 }
 
 @Composable
-fun BottomNavigationBar(navController: NavHostController, items: List<BottomNavigationScreens>){
+fun BottomNavigationBar(navController: NavHostController, items: List<BottomNavigationScreens>) {
 
     val navBackStackEntry by navController.currentBackStackEntryAsState()
     val currentRoute = navBackStackEntry?.destination?.route
@@ -85,24 +97,25 @@ fun BottomNavigationBar(navController: NavHostController, items: List<BottomNavi
 
     if (currentRoute in routes) {
         val currentSection = sections.first { it.route == currentRoute }
-        BottomAppBar(backgroundColor = primaryLightColor,
+        BottomAppBar(
+            backgroundColor = primaryLightColor,
             contentColor = primaryTextColor,
             elevation = 4.dp
         ) {
-            items.forEach{section ->
+            items.forEach { section ->
                 val selected = section == currentSection
                 BottomNavigationItem(
-                    icon = { Icon(section.icon,"contentDescription") },
+                    icon = { Icon(section.icon, "contentDescription") },
                     label = { Text(stringResource(id = section.resourceId)) },
                     selected = selected,
-                    selectedContentColor = primaryTextColor ,
+                    selectedContentColor = primaryTextColor,
                     unselectedContentColor = primaryDarkColor,
                     alwaysShowLabel = true, // This hides the title for the unselected items
                     onClick = {
                         // This if check gives us a "singleTop" behavior where we do not create a
                         // second instance of the composable if we are already on that destination
                         if (currentRoute != section.route) {
-                            navController.navigate(section.route){
+                            navController.navigate(section.route) {
                                 launchSingleTop = true
                                 restoreState = true
                                 popUpTo(findStartDestination(navController.graph).id) {
@@ -119,12 +132,59 @@ fun BottomNavigationBar(navController: NavHostController, items: List<BottomNavi
 }
 
 
-
-enum class BottomNavigationScreens(val route: String, @StringRes val resourceId: Int, val icon: ImageVector ) {
+enum class BottomNavigationScreens(
+    val route: String,
+    @StringRes val resourceId: Int,
+    val icon: ImageVector
+) {
     CURRENT(MainDestinations.HOME_ROUTE, R.string.Indication, Icons.Filled.Home),
     TABLE(MainDestinations.TABLE_ROUTE, R.string.Table, Icons.Filled.List),
     CHART(MainDestinations.CHART_ROUTE, R.string.Chart, Icons.Filled.DateRange)
 }
 
+@Preview("WeatherApp")
+@Preview("WeatherApp. Big font",fontScale = 1.4f)
+@Preview("WeatherApp. Small font",fontScale = 0.8f)
+@Composable
+fun WeatherAppPreview() {
 
+    val bottomItems = listOf(
+        BottomNavigationScreens.CURRENT,
+        BottomNavigationScreens.TABLE,
+        BottomNavigationScreens.CHART
+    )
+    Scaffold(
+        scaffoldState = rememberScaffoldState(),
+        bottomBar = {
+            BottomAppBar(
+                backgroundColor = primaryLightColor,
+                contentColor = primaryLightColor
+            ) {
+                bottomItems.forEach { item ->
+                    val selected = stringResource(id = item.resourceId) == "Home"
+                    BottomNavigationItem(
+                        selected = selected,
+                        onClick = { },
+                        icon = { Icon(imageVector = item.icon, contentDescription = item.name) },
+                        label = { Text(stringResource(id = item.resourceId)) },
+                        selectedContentColor = primaryTextColor,
+                        unselectedContentColor = primaryDarkColor
+                    )
+                }
+
+            }
+        }
+    ) {
+        Scaffold(modifier = Modifier.padding(it), backgroundColor = primaryDarkColor) {
+            Box(
+                contentAlignment = Alignment.Center,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .fillMaxHeight()
+            ) {
+                Text(text = "Fragment", fontSize = 50.sp)
+            }
+        }
+    }
+}
