+++ app/src/main/java/ru/voodster/composeweather/weatherapi/WeatherModel.kt
@@ -4,7 +4,6 @@ import com.google.gson.annotations.SerializedName
 import java.text.SimpleDateFormat
 import java.util.*
 
-
 data class WeatherModel(
 
     @SerializedName("Date") var date: Int = 0,
@@ -14,7 +13,6 @@ data class WeatherModel(
     @SerializedName("Press") var press: Int = 0,
     @SerializedName("Temp") var temp: Int = 0
 ) {
-
     /**
      * Str Temp
      *
@@ -52,5 +50,27 @@ data class WeatherModel(
     fun strDayOfMonth(): String =
         SimpleDateFormat("dd/MM", Locale.ROOT)
             .format(Date(date.toLong().times(1000)))
+
+    /**
+     * Fake weather
+     *
+     * @param dateOffset - in minutes
+     * @return
+     */
+    fun fakeWeather(dateOffset:Int):WeatherModel=
+        WeatherModel(Calendar.getInstance().timeInMillis.minus(dateOffset.times(60000)).div(1000).toInt(),
+            (100..250).random(),
+            (10..95).random(),
+            0,
+            (730..785).random(),
+            (-200..300).random())
+
+    fun fakeWeather():WeatherModel=
+        WeatherModel(Calendar.getInstance().timeInMillis.div(1000).toInt(),
+            (100..250).random(),
+            (10..95).random(),
+            0,
+            (730..785).random(),
+            (-200..300).random())
 
 }
